+++ src/cmp/Storage.hpp
@@ -42,7 +42,7 @@ struct Storage {
 
 	BulletData& createBulletData(EntityType e_type, std::size_t e_ID, float speed, bool type, float damage);
 
-	CharacterData& createCharacterData(EntityType e_type, std::size_t e_ID, bool mode, int health, float attackDamage, float attackingCooldown);
+	CharacterData& createCharacterData(EntityType e_type, std::size_t e_ID, bool mode, float health, float attackDamage, float attackingCooldown);
 
 	[[nodiscard]] const vector<AI>&   getAIComponents()   const { return   AIComponents; }
 	[[nodiscard]] 		vector<AI>&   getAIComponents() 	   	{ return   AIComponents; }
